% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frulhns-package.r
\docType{package}
\name{frulhns-package}
\alias{frulhns-package}
\alias{_PACKAGE}
\alias{frulhns}
\title{frulhns: Data package for Kohl, Ostrovsky et al 10.1016/j.cell.2013.11.025}
\description{
Traced neurons and summary statistical data to accompany:
    A bidirectional circuit switch reroutes pheromone signals in male and female
    brains Cell, Volume 155, Issue 7, 1610-1623, 19 December 2013
    doi:10.1016/j.cell.2013.11.025 Johannes Kohl*, Aaron D. Ostrovsky*, Shahar
    Frechter, Gregory S.X.E. Jefferis.
}
\keyword{internal}
