\name{poissonTestOdours}
\alias{poissonTestOdours}
\title{Carry out poisson test on absolute number of spikes in odour response}
\usage{
  poissonTestOdours(x, odours = colnames(x),
    maxtrials = NA, ...)
}
\description{
  Carry out poisson test on absolute number of spikes in
  odour response
}

