% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frulhns-data.r
\docType{data}
\name{allspikes}
\alias{allspikes}
\title{List containing spiking data for all 288 lateral horn neurons in the paper}
\description{
Each element in the list is a \code{spiketimes} object containing all 
the spiking responses to a single recorded neuron. The spikes were extracted 
in Igor Pro + Neuromatic as described in the Supplemental Experimental 
Procedures of the paper. They were then read into R and post-processed using 
the gphys package, including associating the stimulus information (odour
names, stimulus duration, odour delivery device channel etc.)
}
\examples{
\donttest{
# a cell that is highly selective for cVA pheromone
gphys::PlotRasterFromSweeps(allspikes[[1]])
# note that the external TTL puslse occured at 2000 ms with 0 delay after that
head(attr(allspikes[[1]],'oddconf'))
}
}
\seealso{
\code{\link[gphys]{spiketimes}, \link{fixedUseful}}
}
