% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frulhns-data.r
\docType{data}
\name{fixedUseful}
\alias{fixedUseful}
\title{Dataframe combining cell metadata and odours response summaries for 254 LHNs}
\description{
Each row is a neuron whose electrophysiological data met quality thesholds. 
In contrast to \code{\link{allspikes}} these data do not include cells from 
direct glomerular stimulation experiments.
}
\seealso{
\code{\link{asr}, \link{allspikes}}
}
