\docType{data}
\name{IS2Surf}
\alias{IS2Surf}
\title{Surface model of the IS2 Template brain used for image registration}
\description{
  Surface model of the IS2 Template brain used for image
  registration
}
\examples{
\dontrun{
# Depends on AnalysisSuite
plot3dsurface(IS2Surf)
}
}
\references{
  Cachero S., Ostrovsky A.D., Yu J.Y., Dickson B.J., and
  Jefferis G.S.X.E. (2010). Sexual dimorphism in the fly
  brain. Curr Biol 20 (18), 1589-601.
}
\keyword{datasets}

