% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikefunctions.r
\name{poissonTestOdours}
\alias{poissonTestOdours}
\title{Carry out poisson test on absolute number of spikes in odour response}
\usage{
poissonTestOdours(x, odours = colnames(x), maxtrials = NA, ...)
}
\arguments{
\item{x}{An m x n odour response matrix for m trials of n odours}

\item{odours}{Names of odours (defaults to \code{colnames(x)})}

\item{maxtrials}{Maximum number of trials to consider (default => all)}

\item{...}{Additional parameters passed to \code{poisson.test}}
}
\description{
Carry out poisson test on absolute number of spikes in odour response
}
\details{
It is assumed that x contains one column called \code{"blank"},
  against which the other columns will all be tested.
}
\seealso{
\code{\link{poisson.test}}
}
